Huge refactor with port to ES6
