+++ README.md
@@ -1,0 +1,3 @@
+# DivinePride Bot
+
+A bot that consumes DivinePride's API and answers in Discord and Twitch
