+++ hub.js
@@ -1,6 +1,6 @@
-import Client from './clients/client'
-import DiscordHandler from './clients/discord'
-import TwitchHandler from './clients/twitch'
+import Client from './clients/client.js'
+import DiscordHandler from './clients/discord.js'
+import TwitchHandler from './clients/twitch.js'
 
 /**
  * Config object
@@ -14,7 +14,13 @@ import TwitchHandler from './clients/twitch'
  *  api_key: 'foo'
  * }
  */
-module.exports = function(config) {
-  Client(config.api_key, new DiscordHandler(config.discord_token))
-  Client(config.api_key, new TwitchHandler(config.twitch))
+export default (config) => {
+  if (typeof config.api_key)
+    throw new Error('api_key is empty and is required to consume DivinePride data')
+
+  if (typeof config.discord_token !== 'undefined')
+    Client(config.api_key, new DiscordHandler(config.discord_token))
+
+  if (typeof config.twitch !== 'undefined')
+    Client(config.api_key, new TwitchHandler(config.twitch))
 }
