+++ package.json
@@ -5,8 +5,12 @@
   "repository": "git@github.com:sergiopvilar/divinepride.git",
   "author": "Sergio Vilar <sergiovilar.r@gmail.com>",
   "license": "MIT",
+  "type": "module",
+  "engines": {
+    "node": "12.x"
+  },
   "scripts": {
-    "start": "node index.js"
+    "start": "node --experimental-modules index.js"
   },
   "dependencies": {
     "chatbot-commander": "^0.0.3",
